debugged some compiler errors

+++ kernel/network/protocols/ip.h
@@ -1,7 +1,8 @@
 #pragma once
+#include "stdint.h"
 
 typedef struct {
 
 } IpHeaders;
 
-void ipRecv(void* data, size_t size);
+void ipRecv(void* data, uint32_t size);
